Removed target folder to reduce makefile amount

+++ rust-book/chapter-2/src/main.rs
@@ -16,7 +16,7 @@ fn main(){
             .read_line(&mut guess)
             .expect("Failed to read line.");
 
-        let guess: u32 = guess.trim().parse().expect("Please type a number!");
+        let guess: u64 = guess.trim().parse().expect("Please type a number!");
 
         match guess.cmp(&secret_number) {
             Ordering::Less => println!("Too small!"),
